Rename Autocomplete to AssigneeSelector
